Delete appendix on installing packages - duplicate (#76)
We don't need this appendix on installing packages, now we have the RStudio and Jupyterlab "Package management" section

We lose the info on installing from GitHub with install.packages, but I'd rather not encourage using install.packages at all. People can look this up online if they really want it.