Split out tools into their own sections
This allows JupyterLab, RStudio and Airflow to have much more in-depth sections & headers for each of the sections allowing users to navigate easier.

Added an upgrading guide for RStudio that is currently in draft until we finalise which version we are upgrading to

Added Renv to the list of supported package management tools.
We should consider dropping Conda below renv in the future, as it will be our preferred solution, but we still need to support Conda for users that have not upgraded.