Update get-started.md with SSO auth for SSH keys (#468)
We just had a new starter who followed this guidance but couldn't clone any repos.  It turned out his SSH key wasn't authorised for use with the moj-analytical-services org, so I assume this is a new step everyone now needs given the recent changes to sign-on?

This adds the new step to the quickstart to stop anyone else tripping up on this.  Does this work?  I am not an expert in the new auth process!